Node implementation

+++ common/src/main/java/com/shumencoin/beans/helper/TransactionDataHash.java
@@ -3,8 +3,8 @@ package com.shumencoin.beans.helper;
 public class TransactionDataHash {
 	public String from;
 	public String to;
-	public Integer value;
-	public Integer fee;
+	public long value;
+	public long fee;
 	public String data;
 	public String senderPubKey;
 	public String dateCreated;
